#50 some code cleanup
