+++ App/src/pages/Login.tsx
@@ -18,12 +18,11 @@ import logo from "../assets/logo.png";
 import { useForm } from "react-hook-form";
 import {
     AuthenticationResponse,
-    DecodedToken,
+    decodeTokenToUser,
     login,
     LoginDTO,
     LoginProps,
 } from "../services/auth";
-import jwt_decode from "jwt-decode";
 import { User } from "../services/user";
 import { Link, useNavigate } from "react-router-dom";
 import { useState } from "react";
@@ -42,24 +41,8 @@ export default function Login({ setUser }: LoginProps) {
         login(data)
             .then(
                 (res: AuthenticationResponse) => {
-                    console.log(res.data.accessToken);
-                    //decode access token
-                    //save user in context
-                    //redirect to home page
-                    //code:
                     if (res.data.accessToken != null) {
-                        const decodedToken = jwt_decode<DecodedToken>(res.data.accessToken);
-
-                        const user: User = {
-                            firstName: decodedToken.fName,
-                            lastName: decodedToken.lName,
-                            email: decodedToken.email,
-                            roles: decodedToken.roles,
-                            issuedAtTimestamp: decodedToken.iat,
-                            expiresAtTimestamp: decodedToken.exp,
-                            issuer: decodedToken.iss,
-                            audience: decodedToken.aud,
-                        };
+                        const user: User = decodeTokenToUser(res.data.accessToken);
                         setUser(user);
                         navigate("/");
                     }
