+++ App/src/pages/Register.tsx
@@ -16,16 +16,15 @@ import {
 } from "@chakra-ui/react";
 import logo from "../assets/logo.png";
 import { useForm } from "react-hook-form";
-import jwt_decode from "jwt-decode";
 import { User } from "../services/user";
 import { Link, useNavigate } from "react-router-dom";
 import { useState } from "react";
 import {
     AuthenticationResponse,
-    DecodedToken,
     LoginProps,
     RegisterDTO,
     RegisterUser,
+    decodeTokenToUser,
 } from "../services/auth";
 
 export default function Register({ setUser }: LoginProps) {
@@ -42,25 +41,8 @@ export default function Register({ setUser }: LoginProps) {
         RegisterUser(data)
             .then(
                 (res: AuthenticationResponse) => {
-                    console.log(res.data.accessToken);
-                    //decode access token
-                    //save access token in local storage
-                    //save user in context
-                    //redirect to home page
-                    //code:
                     if (res.data.accessToken != null) {
-                        const decodedToken = jwt_decode<DecodedToken>(res.data.accessToken);
-                        console.log(decodedToken);
-                        const user: User = {
-                            firstName: decodedToken.fName,
-                            lastName: decodedToken.lName,
-                            email: decodedToken.email,
-                            roles: decodedToken.roles,
-                            issuedAtTimestamp: decodedToken.iat,
-                            expiresAtTimestamp: decodedToken.exp,
-                            issuer: decodedToken.iss,
-                            audience: decodedToken.aud,
-                        };
+                        const user: User = decodeTokenToUser(res.data.accessToken);
                         setUser(user);
                         navigate("/");
                     }
